// packages/calcite-components/src/components/input/interfaces.ts
export type InputPlacement = "vertical" | "horizontal" | "none";
export type NumberNudgeDirection = "up" | "down";
export type SetValueOrigin = "initial" | "connected" | "user" | "reset" | "direct";
export type TextType =
  | "text"
  | "textarea"
  | "email"
  | "password"
  | "tel"
  | "number"
  | "search"
  | "file"
  | "time"
  | "date";